requirements.txt
chainlit
google-generativeai
faiss-cpu
PyPDF2
python-dotenv
langchain
langchain-community
langchain-huggingface
torch
sentence-transformers
reportlab